requirements.txt
aiogram==3.*
aiohttp
python-dotenv
matplotlib
requests
APScheduler